✨ feat: 코드리뷰 일부 반영/ 반응형 추가
