--- .vite/deps/_metadata.json
@@ -1,8 +1,0 @@
-{
-  "hash": "3c1fd73e",
-  "configHash": "f3f208b2",
-  "lockfileHash": "e953eb3b",
-  "browserHash": "29cde2fb",
-  "optimized": {},
-  "chunks": {}
-}
