--- src/app/(service)/groups/page.tsx
@@ -1,7 +1,0 @@
-/**
- * 팀 페이지를 구성하는 파일입니다.
- */
-
-export default function TeamPage() {
-  return <main>팀 페이지</main>;
-}
