--- src/components/common/modal/AddModal.tsx
@@ -1,11 +1,0 @@
-/**
- * 생성, 추가, 초대 등 긍정 액션에 사용하는 공용 모달입니다.
- */
-
-import ModalFrame from '@/components/common/modal/ModalFrame';
-
-export default function AddModal() {
-  // TODO: ModalFrame을 기반으로 파란색 주요 버튼을 가진 모달을 구현합니다.
-  // TODO: 목록 추가, 할 일 추가, 멤버 초대처럼 확인 액션이 필요한 곳에서 재사용합니다.
-  return <ModalFrame>추가 모달</ModalFrame>;
-}
