--- src/components/common/modal/DeleteModal.tsx
@@ -1,11 +1,0 @@
-/**
- * 삭제, 탈퇴 등 위험 액션에 사용하는 공용 모달입니다.
- */
-
-import ModalFrame from '@/components/common/modal/ModalFrame';
-
-export default function DeleteModal() {
-  // TODO: ModalFrame을 기반으로 빨간색 주요 버튼을 가진 모달을 구현합니다.
-  // TODO: 팀 삭제, 할 일 삭제, 댓글 삭제, 회원 탈퇴처럼 되돌리기 어려운 액션에 사용합니다.
-  return <ModalFrame>삭제 모달</ModalFrame>;
-}
