+++ src/components/common/modal/components/ModalFrame.tsx
@@ -5,7 +5,7 @@
  * - hasIcon (boolean): 아이콘 표시 여부 (true 시 느낌표 아이콘)
  * - hasCloseButton (boolean): 닫기 버튼 표시 여부
  * - title (string): 타이틀
- * - subTitle (string): 서브 타이틀
+ * - subTitle (string): 서브 타이틀 (멤버 몇명에 사용)
  * - description (string): 기본 문구 (진한 텍스트)
  * - subDescription (string): 서브 문구 (회색 텍스트)
  * - onClose (() => void): 닫기 버튼 클릭 핸들러
@@ -19,27 +19,10 @@
  */
 
 'use client';
-import type { ReactNode } from 'react';
 import Image from 'next/image';
 import { icCloseMedium, icAlertCircleLarge } from '@/assets/index';
+import { ModalFrameProps } from '../types';
 
-type ModalFrameProps = {
-  children?: ReactNode;
-  hasIcon?: boolean;
-  hasCloseButton?: boolean;
-  title?: string;
-  subTitle?: string;
-  description?: string;
-  subDescription?: string;
-  onClose?: () => void;
-  lineButtonText?: string;
-  onLineButtonClick?: () => void;
-  primaryButtonText?: string;
-  onPrimaryButtonClick?: () => void;
-  subButtonText?: string;
-  onSubButtonClick?: () => void;
-  isButtonAlign?: boolean;
-};
 export default function ModalFrame({
   children,
   hasIcon,
@@ -59,11 +42,12 @@ export default function ModalFrame({
 }: ModalFrameProps) {
   return (
     <div
-      className="fixed top-0 left-0 right-0 bottom-0 flex justify-center items-center bg-black/60 z-9999 px-4"
+      className="fixed top-0 left-0 right-0 bottom-0 flex 
+      justify-center items-end-safe md:justify-center md:items-center md:px-4 bg-black/60 z-9999"
       onClick={onClose}
     >
       <div
-        className="max-w-sm bg-white rounded-2xl p-6 pt-10 min-w-80 w-full relative text-center flex flex-col gap-4 items-center"
+        className="w-full bg-white rounded-tl-xl rounded-tr-xl  p-6 pt-10 min-w-80 relative text-center flex flex-col gap-4 items-center md:max-w-sm md:rounded-br-xl md:rounded-bl-xl"
         onClick={(e) => e.stopPropagation()}
       >
         {hasCloseButton && (
@@ -100,11 +84,11 @@ export default function ModalFrame({
         {children && <div>{children}</div>}
 
         <div
-          className={`flex gap-2 mt-2 w-full ${isButtonAlign ? 'flex-col px-4' : 'flex-row px-0'} `}
+          className={`flex gap-2 mt-2 w-full ${isButtonAlign ? 'flex-col px-4' : 'flex-row px-0'} max-w-80`}
         >
           {lineButtonText && (
             <button
-              className="border border-border-secondary rounded-xl px-4 py-2 w-full text-text-default font-medium"
+              className="border border-border-secondary rounded-xl px-4 py-2 w-full text-text-default font-medium "
               onClick={onLineButtonClick}
             >
               {lineButtonText}
